+++ custom_components/wyoming_dotstts/const.py
@@ -1,0 +1,9 @@
+"""Constants for the Wyoming dots.tts integration."""
+
+DOMAIN = "wyoming_dotstts"
+
+CONF_HOST = "host"
+CONF_PORT = "port"
+CONF_API_TOKEN = "api_token"
+
+DEFAULT_PORT = 8180